[chore] Expose namedpipe config on all platforms
