+++ pkg/stanza/operator/input/namedpipe/namedpipe_nonlinux.go
@@ -1,0 +1,18 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+//go:build !linux
+// +build !linux
+
+package namedpipe // import "github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/input/namedpipe"
+
+import (
+	"errors"
+
+	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator"
+	"go.uber.org/zap"
+)
+
+func (c *Config) Build(logger *zap.SugaredLogger) (operator.Operator, error) {
+	return nil, errors.New("namedpipe input operator is only supported on linux")
+}
